feat: 탭별 개별 뷰 통합 및 분리
새로운 산책 시작, 루틴 설정, 산책 기록, 프로필 뷰 추가
