+++ mungcourse/ContentView.swift
@@ -13,27 +13,28 @@ let themeColor = Color(hex: "48CF6E")
 struct ContentView: View {
     var body: some View {
         TabView {
+            // 각 탭에 새로 생성한 뷰를 연결합니다.
             HomeView()
                 .tabItem {
                     Label("홈", systemImage: "house.fill")
                 }
 
-            Text("산책 시작 화면 (구현 예정)")
+            StartWalkView() // 분리된 뷰 사용
                 .tabItem {
                     Label("산책 시작", systemImage: "figure.walk")
                 }
 
-            Text("루틴 설정 화면 (구현 예정)")
+            RoutineSettingsView() // 분리된 뷰 사용
                 .tabItem {
                     Label("루틴 설정", systemImage: "gearshape.fill")
                 }
 
-            Text("산책 기록 화면 (구현 예정)")
+            WalkHistoryView() // 분리된 뷰 사용
                 .tabItem {
                     Label("산책 기록", systemImage: "list.bullet")
                 }
 
-            Text("프로필 화면 (구현 예정)")
+            ProfileTabView() // 분리된 뷰 사용
                 .tabItem {
                     Label("프로필", systemImage: "person.fill")
                 }
@@ -42,71 +43,10 @@ struct ContentView: View {
     }
 }
 
-struct HomeView: View {
-    var body: some View {
-        ScrollView { // 내용이 길어질 수 있으므로 ScrollView 사용
-            VStack(spacing: 20) { // 섹션 간 간격 설정
-                ProfileArea()
-                ButtonArea()
-                NearbyTrailsArea()
-                WalkIndexArea()
-                PastRoutesArea()
-                Spacer() // 남은 공간 채우기
-            }
-            .padding() // 전체적인 패딩 추가
-        }
-        .navigationTitle("홈") // 네비게이션 타이틀 설정 (필요시 NavigationView로 감싸야 함)
-    }
-}
-
-// --- Placeholder Views ---
-
-struct ProfileArea: View {
-    var body: some View {
-        Text("프로필 영역")
-            .frame(maxWidth: .infinity, minHeight: 100)
-            .background(Color.gray.opacity(0.2))
-            .cornerRadius(10)
-    }
-}
-
-struct ButtonArea: View {
-    var body: some View {
-        Text("버튼 영역")
-            .frame(maxWidth: .infinity, minHeight: 50)
-            .background(Color.gray.opacity(0.2))
-            .cornerRadius(10)
-    }
-}
-
-struct NearbyTrailsArea: View {
-    var body: some View {
-        Text("주변 산책로 영역")
-            .frame(maxWidth: .infinity, minHeight: 150)
-            .background(Color.gray.opacity(0.2))
-            .cornerRadius(10)
-    }
-}
-
-struct WalkIndexArea: View {
-    var body: some View {
-        Text("산책 지수 영역")
-            .frame(maxWidth: .infinity, minHeight: 100)
-            .background(Color.gray.opacity(0.2))
-            .cornerRadius(10)
-    }
-}
-
-struct PastRoutesArea: View {
-    var body: some View {
-        Text("지난 경로 영역")
-            .frame(maxWidth: .infinity, minHeight: 150)
-            .background(Color.gray.opacity(0.2))
-            .cornerRadius(10)
-    }
-}
+// --- HomeView 및 Placeholder 뷰 정의는 HomeView.swift로 이동되었으므로 제거 ---
 
 // --- Helper Extension for Hex Color ---
+// Color 확장은 다른 곳에서도 사용될 수 있으므로 유지하거나 별도 파일로 분리 가능
 extension Color {
     init(hex: String) {
         let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
