+++ mungcourse/ProfileTabView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct ProfileTabView: View {
+    var body: some View {
+        // TODO: Implement Profile Tab View
+        Text("프로필 화면")
+            .navigationTitle("프로필") // 네비게이션 타이틀 설정
+    }
+}
+
+#Preview {
+    ProfileTabView()
+}
