+++ mungcourse/RoutineSettingsView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct RoutineSettingsView: View {
+    var body: some View {
+        // TODO: Implement Routine Settings View
+        Text("루틴 설정 화면")
+            .navigationTitle("루틴 설정") // 네비게이션 타이틀 설정
+    }
+}
+
+#Preview {
+    RoutineSettingsView()
+}
