+++ mungcourse/StartWalkView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct StartWalkView: View {
+    var body: some View {
+        // TODO: Implement Start Walk View
+        Text("산책 시작 화면")
+            .navigationTitle("산책 시작") // 네비게이션 타이틀 설정
+    }
+}
+
+#Preview {
+    StartWalkView()
+}
