+++ mungcourse/WalkHistoryView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct WalkHistoryView: View {
+    var body: some View {
+        // TODO: Implement Walk History View
+        Text("산책 기록 화면")
+            .navigationTitle("산책 기록") // 네비게이션 타이틀 설정
+    }
+}
+
+#Preview {
+    WalkHistoryView()
+}
